Ajustes controller user e implemetação metodos no controller provider

+++ app/Controllers/Http/User/ServiceController.js
@@ -4,6 +4,7 @@
 /** @typedef {import('@adonisjs/framework/src/Response')} Response */
 /** @typedef {import('@adonisjs/framework/src/View')} View */
 const Service = use('App/Models/Service')
+const User = use('App/Models/User')
 const Transformer = use('App/Transformers/Admin/ServiceTransformer')
 
 /**
@@ -19,20 +20,29 @@ class ServiceController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async index({ request, response, pagination, transform, auth }) {
+  async index({ request, response, pagination, transform }) {
     try {
-      const name = request.input('name')
-      const user = await auth.authenticator('user').getUser()
-      const query = Service.query()
-      if (name) {
-        query.where('name', 'LIKE', `%${name}%`)
+      const { id, token, name } = request.all()
+      let user = await User.findOrFail(id)
+
+      if (user.token == token) {
+        try {
+          const query = Service.query()
+          if (name) {
+            query.where('name', 'LIKE', `%${name}%`)
+          }
+          query.where('user_id', user.id)
+          var services = await query.paginate(pagination.page, pagination.limit)
+          services = await transform.paginate(services, Transformer)
+          return response.send({ success: true, data: services })
+        } catch (error) {
+          return response.send({ success: false, message: 'Falha ao tentar listar serviços' })
+        }
+      } else {
+        return response.send({ success: false, message: 'Falha na autenticação' })
       }
-      query.where('user_id', user.id)
-      var services = await query.paginate(pagination.page, pagination.limit)
-      services = await transform.paginate(services, Transformer)
-      return response.send({ success: true, data: services })
     } catch (error) {
-      return response.send({ success: false, message: 'Falha ao tentar listar serviços' })
+      return response.send({ success: false, message: 'Falha na operação' })
     }
   }
 
@@ -45,13 +55,19 @@ class ServiceController {
    * @param {Response} ctx.response
    */
   async store({ request, response, transform }) {
-    try {
-      const { name, description, value, initial_datetime, final_datetime, status, category_id, user_id, provider_id } = request.all()
-      var service = await Service.create({ name, description, value, initial_datetime, final_datetime, status, category_id, user_id, provider_id })
-      service = await transform.item(service, Transformer)
-      return response.status(201).send({ success: true, data: service })
-    } catch (error) {
-      return response.status(400).send({ success: false, message: 'Erro ao tentar cadastrar serviço' })
+    const { id, token, name, description, value, initial_datetime, final_datetime, status, category_id, provider_id } = request.all()
+    let user = await User.findOrFail(id)
+
+    if (user.token == token) {
+      try {
+        var service = await Service.create({ name, description, value, initial_datetime, final_datetime, status, category_id, user_id: user.id, provider_id })
+        service = await transform.item(service, Transformer)
+        return response.status(201).send({ success: true, data: service })
+      } catch (error) {
+        return response.status(400).send({ success: false, message: 'Erro ao tentar cadastrar serviço' })
+      }
+    } else {
+      return response.send({ success: false, message: 'Falha na autenticação' })
     }
   }
 
@@ -64,14 +80,20 @@ class ServiceController {
    * @param {Response} ctx.response
    * @param {View} ctx.view
    */
-  async show({ params, request, response, view, transform }) {
-    try {
-      const id = request.input('id')
-      var service = await Service.findOrFail(id)
-      service = await transform.item(service, Transformer)
-      return response.send({ success: true, data: service })
-    } catch (error) {
-      return response.send({ success: false, message: 'Falha ao tentar exibir serviço' })
+  async show({ params, request, response, transform }) {
+    const { id, token, service_id } = request.all()
+    let user = await User.findOrFail(id)
+
+    if (user.token == token) {
+      try {
+        var service = await Service.findOrFail(service_id)
+        service = await transform.item(service, Transformer)
+        return response.send({ success: true, data: service })
+      } catch (error) {
+        return response.send({ success: false, message: 'Falha ao tentar exibir serviço' })
+      }
+    } else {
+      return response.send({ success: false, message: 'Falha na autenticação' })
     }
   }
 
@@ -84,15 +106,21 @@ class ServiceController {
    * @param {Response} ctx.response
    */
   async update({ params, request, response, transform }) {
-    try {
-      const { id, name, description, value, initial_datetime, final_datetime, status, category_id, user_id, provider_id } = request.all()
-      var service = await Service.findOrFail(id)
-      service.merge({ name, description, value, initial_datetime, final_datetime, status, category_id, user_id, provider_id })
-      await service.save()
-      service = await transform.item(service, Transformer)
-      return response.send({ success: true, data: service })
-    } catch (error) {
-      return response.send({ success: false, message: 'Falha ao tentar atualizar serviço' })
+    const { id, token, service_id, name, description, value, initial_datetime, final_datetime, status, category_id, user_id, provider_id } = request.all()
+    let user = await User.findOrFail(id)
+
+    if (user.token == token) {
+      try {
+        var service = await Service.findOrFail(service_id)
+        service.merge({ name, description, value, initial_datetime, final_datetime, status, category_id, user_id, provider_id })
+        await service.save()
+        service = await transform.item(service, Transformer)
+        return response.send({ success: true, data: service })
+      } catch (error) {
+        return response.send({ success: false, message: 'Falha ao tentar atualizar serviço' })
+      }
+    } else {
+      return response.send({ success: false, message: 'Falha na autenticação' })
     }
   }
 
@@ -105,13 +133,19 @@ class ServiceController {
    * @param {Response} ctx.response
    */
   async destroy({ params, request, response }) {
-    try {
-      const id = request.input('id')
-      const service = await Service.findOrFail(id)
-      await service.delete()
-      return response.send({ success: true })
-    } catch (error) {
-      return response.status(500).send({ success: false, message: 'Falha ao tentar deletar serviço' })
+    const { id, token, service_id } = request.all()
+    let user = await User.findOrFail(id)
+
+    if (user.token == token) {
+      try {
+        const service = await Service.findOrFail(service_id)
+        await service.delete()
+        return response.send({ success: true })
+      } catch (error) {
+        return response.status(500).send({ success: false, message: 'Falha ao tentar deletar serviço' })
+      }
+    } else {
+      return response.send({ success: false, message: 'Falha na autenticação' })
     }
   }
 }
